requirements.txt
timm==0.3.2
mmcv==1.3.8
transformers==4.10.2
reformer_pytorch==1.4.3
pillow==8.1.0
numpy==1.17
enum34>=1.1.10